Changes following review
