+++ src/BloomBrowserUI/bookEdit/toolbox/decodableReader/decodableReader.htm
@@ -4,7 +4,7 @@
   <body>
     <h3 data-panelId="decodableReaderTool" data-markuptype="1" data-order="10" data-i18n="Toolbox.DecodableReaderTool">Decodable Reader Tool</h3>
     <div data-panelId="decodableReaderTool">
-      <div id="setupStages"><img id="decodable-edit" src="/bloom/images/edit-white.png"/><span class="setup noSelect"><a href="javascript:parent.window.showSetupDialog('stages');" data-i18n="DecodableReaderTool.SetUpStages">Set up Stages</a></span></div>
+      <div id="setupStages"><img id="decodable-edit" src="/bloom/images/edit-white.png"/><span class="setup noSelect"><a href="javascript:window.showSetupDialog('stages');" data-i18n="DecodableReaderTool.SetUpStages">Set up Stages</a></span></div>
       <div class="stageLine clear noSelect"><span id="decStage" class="scroll-button ui-icon ui-icon-triangle-1-w"></span><span class="stageLabel stageLine noSelect"><span data-i18n="DecodableReaderTool.Stage">Stage</span> <span id="stageNumber">1</span> <span data-i18n="DecodableReaderTool.StageOf" class="ofStage">of</span> <span id="numberOfStages" class="ofStage">2</span></span><span id="incStage" class="scroll-button ui-icon ui-icon-triangle-1-e"></span></div>
       <div id="letters-in-this-stage" style="margin-left: 6px" class="section clear"><span data-i18n="DecodableReaderTool.LettersInThisStage">Letters in this stage</span></div>
       <div class="tableHolder clear">
