+++ src/BloomBrowserUI/bookEdit/toolbox/decodableReader/decodableReader.jade
@@ -8,7 +8,7 @@ html
 			#setupStages
 				img#decodable-edit(src='/bloom/images/edit-white.png')
 				span.setup.noSelect
-					a(href="javascript:parent.window.showSetupDialog('stages');", data-i18n='DecodableReaderTool.SetUpStages') Set up Stages
+					a(href="javascript:window.showSetupDialog('stages');", data-i18n='DecodableReaderTool.SetUpStages') Set up Stages
 			.stageLine.clear.noSelect
 				span.scroll-button.ui-icon.ui-icon-triangle-1-w#decStage
 				span.stageLabel.stageLine.noSelect
