+++ src/BloomBrowserUI/bookEdit/toolbox/decodableReader/decodableReader.js
@@ -2,12 +2,12 @@
 var DecodableReaderModel = (function () {
     function DecodableReaderModel() {
     }
-    DecodableReaderModel.prototype.restoreSettings = function (opts) {
-        if (opts['decodableReaderState']) {
+    DecodableReaderModel.prototype.restoreSettings = function (settings) {
+        if (settings['decodableReaderState']) {
             var state = libsynphony.dbGet('drt_state');
             if (!state)
                 state = new DRTState();
-            var decState = opts['decodableReaderState'];
+            var decState = settings['decodableReaderState'];
             if (decState.startsWith("stage:")) {
                 var parts = decState.split(";");
                 state.stage = parseInt(parts[0].substring("stage:".length));
@@ -55,4 +55,130 @@ var DecodableReaderModel = (function () {
     return DecodableReaderModel;
 })();
 tabModels.push(new DecodableReaderModel());
+// "region" ReaderSetup dialog
+function CreateConfigDiv(title) {
+    var dialogContents = $('<div id="synphonyConfig" title="' + title + '"/>').appendTo($(parentDocument()).find("body"));
+    var html = '<iframe id="settings_frame" src="/bloom/bookEdit/toolbox/decodableReader/readerSetup/ReaderSetup.htm" scrolling="no" ' +
+        'style="width: 100%; height: 100%; border-width: 0; margin: 0" ' +
+        'onload="document.getElementById(\'toolbox\').contentWindow.initializeReaderSetupDialog()"></iframe>';
+    dialogContents.append(html);
+    return dialogContents;
+}
+function parentDocument() {
+    return window.parent.document;
+}
+function settingsFrameWindow() {
+    return parentDocument().getElementById('settings_frame').contentWindow;
+}
+function showSetupDialog(showWhat) {
+    var toolbox = window;
+    var lm = toolbox.localizationManager;
+    lm.loadStrings(getSettingsDialogLocalizedStrings(), null, function () {
+        var title;
+        if (showWhat == 'stages')
+            title = lm.getText('ReaderSetup.SetUpDecodableReaderTool', 'Set up Decodable Reader Tool');
+        else
+            title = lm.getText('ReaderSetup.SetUpLeveledReaderTool', 'Set up Leveled Reader Tool');
+        var dialogContents = CreateConfigDiv(title);
+        var h = 580;
+        var w = 720;
+        var size = getAppropriateDialogSize(h, w);
+        h = size[0];
+        w = size[1];
+        toolbox.model.setupType = showWhat;
+        $(dialogContents).dialog({
+            autoOpen: true,
+            modal: true,
+            buttons: {
+                Help: {
+                    // For consistency, I would have made this 'Common.Help', but we already had 'HelpMenu.Help Menu' translated
+                    text: lm.getText('HelpMenu.Help Menu', 'Help'),
+                    class: 'left-button',
+                    click: function () {
+                        settingsFrameWindow().postMessage('Help', '*');
+                    }
+                },
+                OK: {
+                    text: lm.getText('Common.OK', 'OK'),
+                    click: function () {
+                        settingsFrameWindow().postMessage('OK', '*');
+                    }
+                },
+                Cancel: {
+                    text: lm.getText('Common.Cancel', 'Cancel'),
+                    click: function () {
+                        $(this).dialog("close");
+                    }
+                }
+            },
+            close: function () {
+                $(this).remove();
+                fireCSharpEvent('setModalStateEvent', 'false');
+            },
+            open: function () {
+                $('#synphonyConfig').css('overflow', 'hidden');
+                $('button span:contains("Help")').prepend('<i class="fa fa-question-circle"></i> ');
+            },
+            height: h,
+            width: w
+        });
+        fireCSharpEvent('setModalStateEvent', 'true');
+    });
+}
+function getAppropriateDialogSize(preferredHeight, preferredWidth) {
+    var h = preferredHeight;
+    var w = preferredWidth;
+    // This height and width will fit inside the "800 x 600" settings
+    var sw = parentDocument().body.scrollWidth;
+    if (sw < 583) {
+        h = 460;
+        w = 390;
+    }
+    else if ((sw < 723) || (window.parent.innerHeight < 583)) {
+        h = 460;
+        w = 580;
+    }
+    return [h, w];
+}
+function getSettingsDialogLocalizedStrings() {
+    // Without preloading these, they are not available when the dialog is created
+    var pairs = {};
+    pairs['ReaderSetup.SetUpDecodableReaderTool'] = 'Set up Decodable Reader Tool';
+    pairs['ReaderSetup.SetUpLeveledReaderTool'] = 'Set up Leveled Reader Tool';
+    pairs['HelpMenu.Help Menu'] = 'Help';
+    pairs['Common.OK'] = 'OK';
+    pairs['Common.Cancel'] = 'Cancel';
+    return pairs;
+}
+//noinspection JSUnusedGlobalSymbols
+/**
+ * Used by the settings_frame to initialize the setup dialog
+ */
+function initializeReaderSetupDialog() {
+    var model = window.model;
+    var sourceMsg = 'Data\n' + JSON.stringify(model.getSynphony().source);
+    var fontMsg = 'Font\n' + model.fontName;
+    settingsFrameWindow().postMessage(sourceMsg, '*');
+    settingsFrameWindow().postMessage(fontMsg, '*');
+}
+/**
+ * Called by C# after the setup data has been saved, following Save click.
+ */
+function closeSetupDialog() {
+    $(parentDocument()).find('#synphonyConfig').dialog("close");
+}
+/**
+ * Fires an event for C# to handle
+ * @param {String} eventName
+ * @param {String} eventData
+ */
+// Enhance: JT notes that this method pops up from time to time; can we consolidate?
+function fireCSharpEvent(eventName, eventData) {
+    var event = new MessageEvent(eventName, { 'bubbles': true, 'cancelable': true, 'data': eventData });
+    document.dispatchEvent(event);
+    // For when we someday change this file to TypeScript... since the above ctor is not declared anywhere.
+    // Solution III (works)
+    //var event = new (<any>MessageEvent)(eventName, { 'view': window, 'bubbles': true, 'cancelable': true, 'data': eventData });
+}
+// "endregion" ReaderSetup dialog
 //# sourceMappingURL=decodableReader.js.map
