+++ src/BloomBrowserUI/bookEdit/toolbox/decodableReader/readerSetup/ReaderSetup.htm
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+﻿<!DOCTYPE html>
 <html>
   <head>
     <meta charset="UTF-8">
@@ -18,7 +18,7 @@
     <script src="/bloom/lib/jquery.i18n.custom.js"></script>
     <script src="/bloom/lib/long-press/jquery.mousewheel.js"></script>
     <script src="/bloom/lib/long-press/jquery.longpress.js"></script>
-    <script src="/bloom/bookEdit/js/readerSettings.js"></script>
+    <script src="/bloom/bookEdit//toolbox/decodableReader/readerSettings.js"></script>
     <script src="readerSetup.io.js"></script>
     <script src="readerSetup.ui.js"></script>
     <script type="text/javascript">$(function() {$("#dlstabs").tabs();});</script>
