+++ src/BloomBrowserUI/bookEdit/toolbox/decodableReader/readerSetup/ReaderSetup.jade
@@ -18,7 +18,7 @@ html
 		script(src='/bloom/lib/jquery.i18n.custom.js')
 		script(src='/bloom/lib/long-press/jquery.mousewheel.js')
 		script(src='/bloom/lib/long-press/jquery.longpress.js')
-		script(src='/bloom/bookEdit/js/readerSettings.js')
+		script(src='/bloom/bookEdit//toolbox/decodableReader/readerSettings.js')
 		script(src='readerSetup.io.js')
 		script(src='readerSetup.ui.js')
 		script(type='text/javascript') $(function() {$("#dlstabs").tabs();});
