+++ src/BloomBrowserUI/bookEdit/toolbox/decodableReader/readerSetup/readerSetup.io.js
@@ -83,19 +83,19 @@ function loadReaderSetupData(jsonData) {
     });
 }
 function saveClicked() {
+    var toolbox = toolboxWindow();
     // update more words
     if ((document.getElementById('dls_more_words').value !== previousMoreWords)
         || (parseInt($('input[name="words-or-letters"]:checked').val()) != 0)) {
-        var toolbox = toolboxWindow();
         // save the changes and update lists
         saveChangedSettings(function () {
             if (typeof toolbox['readerSampleFilesChanged'] === 'function')
                 toolbox['readerSampleFilesChanged']();
-            parent.window['closeSetupDialog']();
+            toolbox.closeSetupDialog();
         });
     }
     else {
-        saveChangedSettings(parent.window['closeSetupDialog']);
+        saveChangedSettings(toolbox.closeSetupDialog);
     }
 }
 /**
