+++ src/BloomBrowserUI/bookEdit/toolbox/leveledReader/leveledReader.htm
@@ -4,7 +4,7 @@
   <body>
     <h3 data-panelId="leveledReaderTool" data-markuptype="2" data-order="20" data-i18n="Toolbox.LeveledReaderTool">Leveled Reader Tool</h3>
     <div data-panelId="leveledReaderTool">
-      <div id="setupStages"><img id="leveled-edit" src="/bloom/images/edit-white.png"/><span class="setup noSelect"><a href="javascript:parent.window.showSetupDialog('levels');" data-i18n="LeveledReaderTool.SetUpLevels">Set up Levels</a></span></div>
+      <div id="setupStages"><img id="leveled-edit" src="/bloom/images/edit-white.png"/><span class="setup noSelect"><a href="javascript:window.showSetupDialog('levels');" data-i18n="LeveledReaderTool.SetUpLevels">Set up Levels</a></span></div>
       <div class="stageLine clear noSelect"><span id="decLevel" class="scroll-button ui-icon ui-icon-triangle-1-w"></span><span class="stageLabel stageLine noSelect"><span data-i18n="LeveledReaderTool.Level">Level </span><span id="levelNumber">1</span><span data-i18n="LeveledReaderTool.LevelOf" class="ofStage">&nbsp;of&nbsp;</span><span id="numberOfLevels" class="ofStage">2</span></span><span id="incLevel" class="scroll-button ui-icon ui-icon-triangle-1-e"></span></div>
       <table style="margin-left: 6px" class="statistics clear">
         <tr>
