+++ src/BloomBrowserUI/bookEdit/toolbox/leveledReader/leveledReader.jade
@@ -8,7 +8,7 @@ html
 			#setupStages
 				img#leveled-edit(src='/bloom/images/edit-white.png')
 				span.setup.noSelect
-					a(href="javascript:parent.window.showSetupDialog('levels');", data-i18n='LeveledReaderTool.SetUpLevels') Set up Levels
+					a(href="javascript:window.showSetupDialog('levels');", data-i18n='LeveledReaderTool.SetUpLevels') Set up Levels
 			.stageLine.clear.noSelect
 				span.scroll-button.ui-icon.ui-icon-triangle-1-w#decLevel
 				span.stageLabel.stageLine.noSelect
